iop: Add pointer to RPC attributes in iop_iface_t
This is done in the same fashion as field attributes wrt to
iop_struct_t. This is required in order to properly handle RPC alias
in webservices and for easier later evolution of the code. RPC
attributes are handled in iop_iface_t in order to avoid breaking ABI.
The size of iop_iface_t can be changed without breaking ABI but not
iop_rpc_t.

Refs: #11378
Change-Id: I399acff9e6e93f00d658fbb5468315942efe589b
Progress: 65%
Signed-off-by: Nicolas Barbey <nicolas.barbey@intersec.com>