iop-xml-wsdl: Factorize loop on rpc
This is to ease the introduction of RPC aliases in WSDL generation.
In iop-xml-wsdl, one loop multiple times over all interfaces and RPCs.
To avoid code duplication, this is done in a separate function
iop_xwsdl_rpc_loop which takes a block argument being the body of the
loop. This new function iop_xwsdl_rpc_loop is to be modified to
account for RPC aliases in the next patch.

Refs: #11378
Change-Id: I1ec20c01388b774eafee65bbd9f17ea13974184e
Progress: 95%