time: rename into datetime.
Rename the whole time module from lib-common into datetime. That way, we
won't have a "time.h" anymore. This was an issue because the naming
conflicts with the standard <time.h> header. When building mmsx, this is
no real issue, but whenever we want to build the lib-common as a
standalone module this starts being a source of headaches because it's
quite hard to ensure we're going to include the right file.

Change-Id: Id2c815de794124d3b7e557fd71685aa554af0a11
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>