datetime: add time parser.
That function supports three format of timestamps:
* RFC822: [Day, ]D month Year hh:mm:ss[ tz]
* ISO8601: YYYY-MM-DDThh:mm:ss[tz]
* Unix timestamp

The function is case-insensitive and supports most format of timezones:
* literal name (DST, CST, GMT, UTC, Z, ...)
* +/-hh:mm
* +/-hh
* +/-hhmm

In case of missing timezone, the parsing use the timezone of the machine
it runs on.

time_parse_iso8601 has been adapted in order to be case-insensitive, this
may have some side effect on existing callers.

That function is adapted from cdr-parser.c in LMS. The main differences
are the fact it uses time_parse_iso8601 for ISO parsing instead of
strptime and supports more formats for timezones.

Change-Id: Ic40390b150616818bf7aa2fcc38010fbd84557c7
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>