jshint: adding target to check javascript files
This target allows developers to simply run:

    make jshint

in any directory where there is a Makefile to check javascript files according
to the .jshintrc under version control in tools repository.

This patch assumes:

- nodejs has been installed
- npm (NodeJS Package Manager) has been installed
- jshint v1.1.0 (latest npm package) has been installed
- The coding rules .jshintrc file is located in either $HOME/ or somewhere from
the location the target is executed all the way to root. If this file is not
found, jshint will run with its default settings, making the target useless.

Refs: #12961
Change-Id: I8a994d5b14683a33aa6e5b7e62b8c61ef5af27d4