el: makes epoll_wait() return immediately when there's pending signals.
When a signal occurs, a bit is set in _G.gotsigs and we wait for the
completion of el_loop_fds to call the corresponding callbacks. When the
signal occurs during the epoll_wait, this works well since epoll get
interrupted, but when it occurs before the call to epoll_wait, we have to
wait for the completion of the timeout (which may be as large as 59s if no
timers are registered).

This is mainly an issue when using the thr-jobs that use the SIGCHLD to
notify the main thread it has pending jobs. If the signal is received at
the wrong moment, the job may wait for 59s before being executed which may
be really problematic.

The patch reduce the timeout to 0 (dont wait for event on fd, only returns
already fired ones) if a signal processing is pending. It only mitigates
the whole issue since a signal handling can still occurs after the
evaluation of _G.gotsigs and before the call to epoll_wait.

In my stress tests, the issue occurs approximatively 4 times lesser than
previously.

Change-Id: I6d4ee4d0ffbddd5a9cfeac2e2733abda1c261b69
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>