thr-job: use a fd-based event to wake up the main thread.
Previous implementation used a signal, but as explained in 4dc6452, this
is racy and in some situation you'll have to wait for a whole loop of the
event loop for your posted event to be processed.

In order to avoid that issue, we now use a fd (plugged in the event-loop)
to wake the main thread up. This is not racy since epoll_wait() will stop
as soon as an event has been fired (whether it's while epoll_wait() is
running or not).

The implementation use 2 solutions, depending of the system on which it is
*compiled*:
* when the compilation knows about eventfd, then eventfd() is used. It's a
 very fast and lightweight solution.
* when eventfd fails (syscall not implemented?) or when it is not
 available at compile time, a pipe is used.

When we want to wake the main thread up, we write some data on the fd.

Experimentally, on my machine, this totally resolve the issue of some
event being executed after "59s". It takes 7 micro seconds for the eventfd
to wake the main thread, 11 micro seconds for the pipe (12 micro seconds
for the signal).

Change-Id: Icac987ccbe1b31a86518ea5e6fb631975e9c289c
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>