thr-job: fix main thread wake up code when not using el_loop.
In case we were posting jobs on the main queue and not using the el_loop
AND we used the pipe() variant of the wakening code, the pipe get filled
with data but was never emptied. As a consequence when the pipe was full,
xwrite() waited infinitely since nobody on the other side flushed the
content of the pipe.

Use a write instead of xwrite, since we don't care about the content of
the pipe, we just want it to contain some data, if the call fails, this
just means somebody else already tried to wake the main thread up.

Change-Id: Iab3fbaaefe2efff33fa068cbed065443396c8e9d
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>