iop[http]: fix RPC alias callbacks creation in __ichttp_register
Currently when we register an RPC alias we re-use the same
http-iop-callback and we overwrite it with the alias "names". In addition
we also duplicate it one time instead of two times.

We obviously need one callback object per alias. The patch is easy because
we do not allow to unregister a callback, then the only real owner of them
it the trigger hastable which destroys them in its qm_deep_wipe(). So, we
only have to register a clone of the main callback with the alias "names"
and forget about it.

Refs: #11378
Change-Id: If18c4b7e7ff69acc8fcb5bca5be980d3418e6a1f