str-stream: add assertions in unsafe API calls.
Almost every single API of pstream_t comes in to flavors: safe and unsafe.
Currently, there's absolutely no way to ensure unsafe API is used
properly, so this commit add assertions in the unsafe flavor. This will
slow down the code when assertions are enabled, however this will ensure
pstreams are always properly used.

The assertion that is performed is the exact same as the one performed in
the safe version.

If you code get broken after that commit, this means it does something
wrong.

Change-Id: Ib00a96c441afbc235727b88405191d7bac8c1903
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>