make bsf/bsr valgrind friendly
bsf/bsr aren't very friendly with valgrind since we access to (byte *) as
(uint64_t *) because we know that we will never use what in have read in
the possible undefined part of the pointed buffer.

To make valgrind happy, we mark the data that we know (ok… hope…) we will
not access as defined.

Change-Id: I4ad7d1f1afdfde9f70e12799fe107476a7926fe4