thr-jobs: fix race in queue deletion.
Previously, the queue was deleted when, after draining the mpsc, the
closing flag was set. However, there is a race: the closing flag is set
before adding the finalizing job in the queue, as a consequence, the mpsc
can be drained while that job is not yet posted but the flag is already
set. This leads to a double free since a new drain will occur and the
deletion will be performed a second time.

The fix ensure we won't delete the queue before reaching the finalizing
job. As a consequence, there's no need to have a closing flag anymore,
that flag would only allow early (but totally random, since based on the
content of a dangling pointer) detection of double deletions.

Closes: #13870
Change-Id: Ica50d6f3c4ca582371f6f6ab1e55ce9648bc8702
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>