Build: simplifications and fixes for clang-analyzer.
Currently, the loaded cflags are from cc-$(CC_BASE)-flags, that means that
if we use clang/clang-analyzer, the loaded cflags are clang's flags. This
means that we should try to load CLANGFLAGS in that case since they are
not defined.

This drops the support the of COMPILER variable from the build system,
however this does not hurt since it was unused.

Moreover, ensure that the clang-analyzer is properly detected as a clang
variation so that the generated flags are the good ones.

Change-Id: I1525a042dc1a6c89316897f4cd1abea3950e8287
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>