log: introduce a new logging module inspired by log4* family.
This module provides a hierarchical logging mechanism in which each code
module use its own module that can be configured independently from the
rest of the program.

The documentation is provided.

This initial commit provides the core features and an improved (compared
to previous implementation) console logger. Future commits will provide a
unified configuration interface for IOP and probably some improvements in
the log formatting.

Change-Id: I4825e2838c269967caaf204627fa6721e30c5733
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>