iop: add function to handle integer suffix in iop and json files
For the purpose of #6931 we must be able to handle multiplier suffix on
integer constant when reading an iop file as it is done for json files in
iop-json.c:iop_json_lex_number_extensions().

New functions memtoll_ext, memtoull_ext, strtoll_ext, strtoull_ext have
been put in str-num.c. These functions call strtoll or strtoull and process
integer extension. They are intended to be called by lexer of json or iop
file in order to compute the integer value associated to lexemes build from
digits followed by letters.

Helpers for pstream have been added.

Suffix 'k' was not handled before the patch, and it is not handled by the
patch because it would induce misleading between 1000 or 1024.

When applying suffix, test on multiplication overflow has been improved to
handle properly signed numbers.

If an overflow is due to a suffix we saturate the value, as strtoll() does
when it detects an overflow.

Refs: #6931
Change-Id: I863ecc438ad6b8e0fd92a7770fdb01756c1e4bd2