Build: add AddressSanitizer (ASan) support.
AddressSanitizer is a tool integrated with clang >= 3.1 and gcc >= 4.8
that instruments the code in order to perform memory error detection. Some
of those errors are redundant with valgrind's memcheck checks (invalid
read, invalid writes, ...) but some are not.

For example, ASan will detect the following error:

  int foo(int a)
  {
      int tab[2] = { 1, 2 };

      return tab[a];
  }

  int bar(void)
  {
      return foo(-1);
  }

Contrary to valgrind, ASan kills the program on the first error. Meaning a
buggy program just cannot run. Moreover, the performance hit of ASan is
only 1.5x or 2x while valgrind's impact is 20+x.

More information is available at:
https://code.google.com/p/address-sanitizer/wiki/AddressSanitizer
http://clang.llvm.org/docs/AddressSanitizer.html

The patch do the following:
* Add a new build profile P=asan that instruments the code with clang
* Add helper macros to know if we're using Asan or not
* Fix some error that popped up with ASan (more error are fixed in more recent  branches)

What you must keep in mind:
* code using QPS must define the flag 'handle_segv=0':
   ASAN_OPTIONS=handle_segv=0 lib-common/zchk

Change-Id: I61e6724a0be0f8ab26f755bbcf8c33eef9689691
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>