thr: disable pthread_create interception when running with asan.
This interception created some conflicts in case allocations were
performed in an other thread since asan was not able to perform its own
interception of the call, it couldn't instrument the new thread.

The main consequence is that we're going to have issue when running in
asan with thread created by 3rd party libraries.

Change-Id: Iec2fcef038caa43ec1851bc2643072fd4e87d121
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>