bit-stream: fix possible bad memory access in be decoder.
This was not critical since we have the guarantee the data remains in an
accessible location, however, it could be out-of-buffer (and thus lead to
a crash of Address Sanitizer).

We introduce a new byte pointer in the bit-stream structure in order to be
able to access each byte individually.

There will probably be others like that one in the bit-stream.

Change-Id: Ida24eac1ed79af699c64df6bf377d8582848c984
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>