log: ensure argument passed to the macro is a logger_t.
We cast the logger in order to ensure it's non-const, but since C's lack
of a proper const_cast<>, this also allowed casting any structure to a
logger. Now ensure we go through a properly typed const logger_t that is
compatible const and non-const variants and then unconst that one into a
usable logger.

Change-Id: Ib542361adcd2f85ea67bb7564ee56c7ebab3cd0f
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>