ichannel: allow to change the delay of the reconnection timer
A reconnection delay of one second was hard-coded in the IChannel library.
But it could be really useful to change this value. For example, if we
want more reactive setup during tests we may set a smaller value. In the
opposite side, we may want a bigger delay on some setup (slow VPN, …).

This patch is almost free because we use a hole on the ichannel_t
structure.

Change-Id: Id46d61b126575c75598661b2a80346eb9e89b01d