mem: fix valgrind instrumentation of core-mem-fifo.
Due to ASan limitations, we must unpoison memory before unmapping it. As a
consequence a call to mem_tool_allow_memory has been added. However that
call used to mark the memory as undefined, but the size of the map is part
of the map itself, which means that we marks the bytes containing the size
as undefined. This lead to a warning since valgrind complains about the
argument passed to the syscall containing uninitialized bytes.

Closes: #14603
Change-Id: I06c8b25d18d2a66ea61ab9b62e2b381723c7c066
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>