mem: ensure released arenas are unpoisoned.
ASan poisoning is only user-managed, there's absolutely no magic within
ASan that would unpoison freed or unmap pages. This was easily noticeable
for the case of munmap since some addresses get rapidly reused leading to
false positives in ASan. However, this also can be triggered with malloc,
but this requires some time because ASan has a quarantine mechanism to
avoid an address to be returned by malloc shortly after it has been freed.

Change-Id: I4ee932e5863f9cac9981b1a5065867fd0891b594
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>