http: properly support 1XX responses from the server.
According to RFC2616, the server can send as much 1XX response as its
wants. Those responses are no real replies to the request and thus they
must not affect the processing pipeline: the currently processed request
remains the same.

Also add the ability to explicitly request an 100 Continue intermediate
reply from the server. This can be done by setting query->expect100cont
and then waiting for the query->on_100cont callback to the called. We hook
on the fd activity watcher to send the body even without any response from
the server.

Test will be added on branch 2011.2.

Closes: #14734
Change-Id: Idb67ab6859327388f5494d9c88d0f70800dee569
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>