el: add a notion of worker.
A worker is a callback that consumes the idle CPU time of the el_loop, it
is called whenever the event loop might block waiting for a system event.
As such, workers can be used to handle heavy (splittable) tasks that would
otherwise require to be scheduled periodically with timers and execution
time limits (in order to avoid blocking the event loop).

A new el_has_pending_events() call is also provided in order to let the
worker knows when the event loop want to take back the control of the
program. The worker also receive a timeout that is mostly a hint of the
time it is allowed to run, and thus that timeout can be used to estimate
the workload the call can process. However, that timeout may be ignored by
the caller since el_has_pending_events() will return true when that
timeout is reached.

Change-Id: Ie7ecbd5c0d9efc7bd8b99ef5bc4106fc5e6e05b0
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>