Build: add a fast-check target.
This is the first step to a transition that will make slow tests the
default of 'make check' while 'make fast-check' will be available to run
only tests that are not considered to be 'slow'.

The second step will come in a few weeks, when most branches will get
updated to have support for make fast-check. Then, make check will be
properly updated.

Change-Id: Ic7c5894b68021aba67fac8b2b5f69efc045c5087
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>