iop: Simplify __ic_forward_reply_to, enhance doc.
This function, introduced in fc6c4b088, was using a hack to retreive the
serialized answer while it was not necessary because we can directly real
the ichannel read buffer (which was not modified for sure if we are always
in the callback scope).

This simplification in ic_read_process_answer is needed to be able to use
a class as a RPC answer/exception (will be done in the next patch).

Refs: #4412 #14914
Change-Id: Ibfffc08c1d041889bd99b45cce23fd4450276654