iop-dso: Fix symbol resolution.
Before this patch, the packages/structures registered in a DSO were the
ones of the main program (if defined in it), and not the ones contained in
the DSO. As an example, the test added was failing.

By calling dlopen with the flag RTLD_DEEPBIND, we are forcing the symbol
resolution to lookup in first in the DSO symbols.

Change-Id: Ia81fa482af63207d91cf96e5806eafe823a8f296