thr-job: fix possible deadlock in thr_queue_sync.
When consuming the content of the queue, we called
thr_job_dequeue/thr_job_steal in order to fill the idle time of the CPU.
However the jobs run by those dequeue/steal may themselves post new
synchronous jobs on the queue, thus waiting for the completion of one of
its callers... which dead locks.

Tests will come on the master branch.

Change-Id: Iaee41d6245542f4c75369c3fe36be5fa30baa173
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>