core.iop: Add LogFileConfiguration and log_file_create_from_iop.
This new class is inspired by the several IOP structure of our products,
which are all defining their own.

This is the same as mcms::LogConf and lms::LogCfg, with a little
difference: maxTotalSize (int, in MB, default value 1024) was replaced by
totalMaxSize (long, in bytes, default value 1024 * 1024). This change was
motivated by the following reasons:

* this was misleading to have a different unit for maxSize and
  maxTotalSize (and MCMS had a bug because of that). This was probably
  made like that because of the log_file_t API...
* this is dangerous to have a size which unit is not byte, because we
  can't use the "K", "M", "G" json notations.
* the renaming was done to avoid this change to be backward compatible,
  and force the clients to change their value.

This new class is meant to replace mcms::LogConf and lms::LogCfg
structures. igloo::LogCfg and mcg::LogFile structures will inherit
from it.

log_file_create_from_iop is inspired of the function mcms_create_log_file.

Refs: #15170
Change-Id: I7ca2518274bb789ae9ad53a9f48e5047a4db49f6