qhash: fix uint32_t overflow on allocation
Even if size is limited to values in the range [0,UINT32_MAX], it often
get multiplied with factors of either 2 or more (for example the size of
the key or the size of the value). As a consequence, with large hashtable
we can trigger 32bit overflows.

In that case, the first overflow was during the allocation since the
values/keys vector get reallocated to values of newsize * v_size (resp.
k_size). This lead to allocation that were far too short (the truncation
of the value to 32bits) and thus to probable crashes at the first access
to the hashtable.

This patches fixes the most important case (resize) and tries to fix all
possible overflow cases in the remaining of the code.

Signed-off-by: Sebastien Boulle <sebastien.boulle@intersec.com>
Change-Id: Iaaaa0c121ec4909f430940bcbfc73217dfd70959