container: introduce qhugehash.
The qhugehash provides a wrapper around the qhashes. It distributes the
content into a predefined number of buckets, each being a regular qhash.

It's API is nearly identical to the one the qhash, except that the access
to the content of the table require the use of accessors (qhh_key_p,
qhm_value_p) since direct access is not possible.

Moreover, the positions returned by the find/put API are 64bits long (it
contains the identifier of the bucket as well as the position of the entry
in that bucket).

The repartition among the buckets is not guaranteed to be homogeneous
since it depends on the hash function, however on a large dataset it
should be sufficiently homogeneous to scale linearly on the number of
buckets.

Change-Id: I49b1cf502922a0b3428b0ffed71fef6264b777e0
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>