iop: ichannel: be more efficient in case of large messages.
When we know we're in the middle of a large message, set the hint of the
read() call to the missing size of the message. As a consequence, we will
issue much fewer system calls.

Moreover, don't go back to the event loop if using a TCP-based channel and
the first read filled it's entire buffer (so we may have remaining data in
the socket buffer) but was not large enough to read an entire message.
This will avoid issuing useless system calls (clock timers, epoll_wait...)
to complete the read.

The function stills returns if at least an entire message has been read,
even if we may not have read the entire socket buffer.

Change-Id: I184911dbf0975592321d4dc77ad07443a99e676d
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>