iop: ichannel: add a raw mode for the callback.
When dealing with RPC that returns a large amount of data, the decoding
may be expensive, and thus it may block the event loop for a few tens of
milliseconds. In case we know our RPC deals with a huge amount of data, we
know have a possibility to ask the ichannel not to unpack the structure
and only provide the raw pstream. In that case, the pointers res and exn
are left NULL and the message contains the structure.

Change-Id: Ie82229009cc7c1f6e7eef2fe02355d484c5c8b67
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>