iop: inline and rewrite integer decoding function.
Profiling showed that the function never got inlined and that as a
consequence it costs quite a bit of memcpy(). Now, we inline the function
so that the length can be resolved at compile-time, and even if it's not
the case we won't issue a generic memcpy() but a simple memory access.

Change-Id: I1df092cd1f78eb1ba1453dcee799129a348c4bc7
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>