iop: add field ifaces_attrs into struct iop_mod_t
This field was missed in ec83fa9, and is required to handle doxygen
comments for fields of a iop_mod_t structure.

This field is a pointer to data of new type iop_mod_iface_attrs_t which is
intended to handle attributes of a field inside a module and which is
different from iop_iface_attrs_t which is intented to handle comments of a
iop_iface_t structure.

Refs: #8530
Change-Id: I7bd919ecf0799161d4afd5b3f5824b6c4f84f124