iop: fix invalid pointer of empty repeated fields leaved by iop_copy()
When iop_copy() duplicates a repeated field, it set to the destination
field the array length and then the pointer of the duplicated data. But
when the array is empty, the destination pointer is undefined. It works
since the new array length is correctly set to zero, but it will be really
better to have a NULL tab-pointer instead of an invalid address.

Change-Id: I6687d09bb58e168d544a9f0e06c0e098f75a59f5