ichannel: handle local ic when querying RPCs
By default pack/unpack is not done for a local ic, when processing query
arguments constraints are checked. Flag force_pack enable packing of a
message of a local ic.

By default duplication of header and argument are not done when processing
a query for a local ic. Flag force_dup enable duplicating header and
argument. Header has const qualifier in rpc implementation arguments, but
it could be overwritten while processing query if it is simple.

Refs: #8650
Change-Id: I1c158c8d3f9f92b540a4c1d5e7e2920310c43649