log: Fix logger_set_level with level to LOG_LEVEL_DEFAULT.
The level given to logger_set_level can be a "C" level (defines in
syslog.h and log.h), or an "IOP" level (enum LogLevel defined in
core.iop, whose values are mapped on the C values except for DEFAULT).

The level stored in a logger_t is interpreted as a "C" level.

The problem is that the "IOP" value LOG_LEVEL_DEFAULT is equal to -2,
which is the "C" value LOG_UNDEFINED, so when calling logger_set_level
with the IOP value LOG_LEVEL_DEFAULT was actually setting the logger level
to LOG_UNDEFINED, which was leading to really strange behaviors.

To fix that:
* if LOG_LEVEL_DEFAULT is given to logger_set_level, set the logger level
  to the C value LOG_DEFAULT (which is LOG_TRACE or LOG_DEBUG depending on
  the build mode).
* change the C value of LOG_UNDEFINED to avoid conflicts.

Change-Id: I4a4b9c0a304015c346aef0fbcc7710f62752fcdd