thr-jobs: retain the syn when posting the job.
This is a follow-up to 9b5194a358 that ensured the thr_syn_t won't get
deleted before job_done() ended. Instead of retaining the syn directly in
job_done(), we now retain it directly in job_prepare(), which should
reduce contention when posting a lot of jobs using the same syn since all
the retain are done on the same thread.

Change-Id: Ife00284f1d5b55f7345bbc351ead681e681be84f
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>