thr-jobs: steal several jobs at once.
The threads now try to steal several jobs at once in order to reduce
contention on the queue of threads that post a lot of jobs. Threads get
organized using a strictly, unidirectional hierarchy so that a thread get a distance to each
other thread. If a thread has a distance of 1 from another one, it can
steal half of its jobs. If a thread has a distance of 2 from another one,
it can steal a quarter of its jobs.

Change-Id: I772fff7ec4538ef36ef16b219f8fb2bfc9a70257
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>