str-l: Fix mmsx c++ files compil; change prototypes.
mmsx compilation of c++ files was broken by 8d7dbe5853c, because the
compiler was looking for a function "strlen(const void*)" which does not
exist.

To fix that, change the prototype of these functions, that were broken
anyway.  We are now consistent with lstr_dups, mp_lstr_dups, lstr_copys
and mp_lstr_copys.

Note that a small patch will be needed in qts to make it compile with this
patch.

Change-Id: Idd132da53d905aa6d868bca1f64e461d21874441