iop: control default values packing with packer flags
This patch introduces a “packer flag” to control whether we want to pack
the IOP default values or to ignore them.

It was a good idea to ignore the default values to save some bandwidth,
but it's not what we want per default. When we pack IOP object in some
storage we want to unpack them with their old default value or any change
in default values will break backward compatibility.

Change-Id: I87be564f292a62acce618a88f3ab2563a14946bc