iop: fix packer when using an already filled “szs” array
The “szs” array isn't owned by the IOP library. It's given by the API user
and could be already initialized (which is exactly what is done by the
IChannels). So you cannot rely on it to know if we are on the first
recursion of iop_bpack_size_flags().

Anyway, the implementation is better this way.

Change-Id: I1372430973a3021800685305bafb11daf8bc79ac