iop: fix json unpacking when last line is a comment
Currently, when parsing a json file that contains a last line that is a
comment and that does not contain a trailing \n, parsing fails because the
last comment is parsed as a part of the json.

This behaviour is due to the ps_shrink that does not do anything if the end
of the file is found for a one line comment.

This patch just ignores the end of the line if the end of file is found in
a one liner comment.

Change-Id: Ieef936832ffca87224efaeb98871d5ea7738d7b7