iop: add ability to get the allocation size of iop_dup/copy
The iop_dup and iop_copy functions now take an optional pointer in which
it stores the size it has allocated.

This is useful to evaluate the memory pressure of the data kept in RAM. In
order no to break the API for now (as long as the lib-common is not
forked), keep a compatibility layer. This will be removed as soon as
possible.

Change-Id: I69d3ef29d67ddd67e1cb6a6990c0ab4f0a7bf74d
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>