file-log: Call cb with LOG_FILE_DELETE *after* file deletion.
Currently, LMS is the only real user of this log_file_event (and platform
is about to use too, for a similar usage): when his callback is called
with LOG_FILE_DELETE, it updates the log file list by calling glob on the
log directory.

Problem: the log file which was about to be deleted was listed by glob, so
considered as still existing until the next callbacl call.

Solution: Call the callback after the file deletion, and not before.
          If unlink fails, print an error and do not call the callback.

Refs: #16971
Change-Id: Id1c170dd8e57810e7cde48b787a157564836bbf4