Build: add the ability to export variables.
The main goal is to be able to export a list of libs and reuse it
everything it is required. The variables to be exported have to be
explicitly listed. Once exported, they must be accessed with a double $
and the path of the makefile in which they are defined:

Here is a small example (supposing the Makefile is in libcommon/):

 _EXPORT = libs

 libs = libcommon.a time-lp-simple.a

 zchk_SOURCES = zchk.c $$(libcommon/libs)

Change-Id: I7f0b669cb905a4047b7fa835c5ab40c391079a7b
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>