http: Do not url-decode the "vars" part of an URL.
The previous code was URL-decoding the whole URL, including the vars part.
The problem is that it's impossible to properly parse the parameters after
is has been escaped.

For example, if the URL is http://127.0.0.1/?cid%3d1%26titi=8,
the vars part was turned into: cid=1&titi=8
So parsing the vars would lead to have two key/values pairs (cid => 1
and titi => 8) while there should only be one (cid=1&titi => 8).

Refs: #17081
Change-Id: I81353f81aeeed0bf72fd5929fc350bcf73c3adaf