httpd: Gently close the connections in httpd_unlisten.
With the previous implementation, when shutting down a binary using a
HTTTd machine, the potential still opened connections were not properly
closed when calling httpd_unlisten.

The consequence was that the associated HTTP triggers were not wiped, so
the httpd_cfg_t was not wiped neither (because of the refcounting).

By closing gently all the connections when calling httpd_unlisten, the
connections are destroyed at the next main loop execution, so the HTTP
server can be fully wiped at shutdown.

Refs: #17081
Change-Id: Ia60e5d6251692e9103be1e52a7e717345d886ae7