asn1: fix undefined value in writer tests.
The compiler is totally free to fill the padding bytes of literal
structures with the content of its choice. As a consequence, the
assignation only guaranteed the defined field were at the provided value
(and the missing fields were at 0), but not that the padding was left
untouched.

As a consequence, the comparison using memcmp a bit later in the code was
not guaranteed to be verified since it also compared the content of the
padding. This led to a test failure when built with ASan.

Change-Id: Id3ac2f4d34e2065dccf16a3823f4f3c767ce1218
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>