iop: removes useless helpers.
The main issue is that the IOP bloated the compilation units with hundreds
of small functions which caused a dramatic increase in compilation time.
This patch remove the helpers that are totally unused. More job will be
done soon to remove more useless helpers.

Note that this patch improves the compilation speed of LMS by
approximatively 25-30% which is *huge*. IMHO we should get rid of most
helpers as soon as possible to keep the compilation speed of the
repository confortable.

Change-Id: I1fd5ea4f6b6f7bf2e2f16a3c7c983c0499cc9456
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>