core-mem-stack: fix performance regressions.
Commit 918a91c700f3 introduced support for alignment in allocations. As a
consequence, the ROUND_UP() were not computed using divisions by power of
2 anymore, leading to divisions and multiplications to be generated.
However, the alignment is always a power of 2, thus the generated code
became uselessly inefficient.

This patch ensure the alignement is always treated as a power of two and
adapt the arithmetic to ensure the compiler don't get fooled by the
abstraction. This improves the performances of the allocator by a factor
of 2 in release build.

Change-Id: I8b46056346b57addf34d1f7d1960887b69b4d202
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>