iop: avoid getting stuck in proxified rpc
If the proxy is up but the server behind is down, the rpc could get
stuck in the proxy, and the client would never receive any answer.

When the query is made from the php, it is quite problematic since the
whole website is blocked.

This happened in MCG using ic_register_proxy() on an unix socket
ichannel. In that case ic_connect() can fail, and we have ic->elh ==
NULL (and ic_is_ready(ic) == false).

Refs: #17173
Change-Id: I28cece11beabafa479a07cb909dbc35d2f6a32ac