thr-job: fix race when stopping the threads.
When stopping the threads, we set the _G.stopping variable that is read by
each thread in its local processing loop. The thread then exits its loop
and enter the thr_info_cleanup() function. That function cleanup the local
thread context and set the 'alive' marker to false.

The main thread waits for the alive marker to be set to false before
entering the global cleanup stuff that will destroy the thread context
array.

In the previous implementation, the alive marker was set to false before
being completely done with the thread local context. As a consequence, the
main thread could start the global cleanup before being sure all the child
thread were actually stopped. This caused a random crash at process
shutdown.

The fix ensure that setting the alive marker to false is the last
operation of the thread cleanup operation.

Change-Id: I60cfff353184d85c302f583de49c52967f72f2e9
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>