z: flush log stream as soon as written
When forking during a steps logs appears several times in capturing them
into a pipe. This is caused by the fact that for each process forked the
buffered stream is flushed at the end of the process.

The fix is to force the flush of the stream as soon as something is
written in it such that when forking stream buffer will be empty.

Change-Id: I7d2f15b9e0fda1cf98bdd152d5c1a51d6424a7b5