fix backtrace() dead-lock when catching aborts
When we catch a program abort we want to dump the backtrace (using
the backtrace() function). But if we catch a signal coming from malloc()
we dead-lock because backtrace() calls an init() function which performs
some malloc() too.

In this patch we fix this bug by invoking the backtrace() function when
the program starts. It's a dirty hack but also a well-known one (see
https://sourceware.org/bugzilla/show_bug.cgi?id=956).

Change-Id: Icfeccf1d93901603faada165793ac85dc32a6b07