iop-rpc: add a ic_hook_ctx_t to store context of a ic query
We will want to use pre/post hooks for ic query to factorize code,
but to do so we will need to be able to pass information inside the
RPC workflow.
So we made a context and its helpers who will be used to pass these
informations (reply send to abort execution of RPC's implementation,
slot of the request and optional data).

TODO:
    - add the post_hook callback inside it to be able to call it from
asynchronous reply.

Refs: #17689
Change-Id: Ia57960bef60fde78994793bd140ad2a213921320