iop-rpc-channel: call pre/post hook before resolving RPC call
When we call a RPC, we initialize the current ic_hook_context with the
slot of the query and the post callback, then we call the pre_hook with
this context.
Then during the ic_reply we will get the context using the slot of the
query we reply to and do the post_hook thanks to it. Then after this if
we are in an asynchronous reply we will delete the context.

Refs: #17689
Change-Id: Ib4838de3d88d060ecbe02a9a7f875778d8bcc08e