qps: improves writing and reading performances.
For pages maps, we now optimize the writing for speed instead of
optimizing for space. As a consequence, the compressor is much more faster
(~3.8s for 256MB of data). This however remains far below what the
harddrive can handle. Additionally, to reduce the number of writes, the
set a buffer of 1MB to the file object.

For handle maps, we avoid performing a single I/O per handle. Instead we
try to identify large holes (defined by an empirical 4K). As a consequence
the new algorithm writes everything excepted holes. Since most handle maps
are dense at the beginning of the map, this allows reaching the maximum
speed of the disk.

Change-Id: Id402d828067e6c0d14a678ae8cbf53e681454137
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>