iop-rpc: store iop_rpc_t inside ic_cb_entry whatever the type
Currently we don't store rpc type for proxified RPC. This make us
unable to know what exception throw inside the pre_hook of a proxified
query. This also prevent us to make some check (like in igloo check if
an application can use this RPC) directly inside the proxy, forcing us
to send the query to the target to know if we should have thrown an
error.
So we extract it from the union and store it inside the ic_cb_entry for
all the ic_register (callback, proxy, dynproxy, webservice, ...).

Refs: #17689
Change-Id: I37533d830f5c2babb91d6f9efff098d066a25cbb